Move Array-Tag classes into single PlcTags file
